# improve performance for merging markers from overrides