# Pull Request Check List

Related to: #9956

<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes.  But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [x] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

This is far from fixing the regression reported in #9956. It "just" improves the performance (compared to Poetry 2.0) significantly.

Using the simplified repro from https://github.com/python-poetry/poetry/issues/9956#issuecomment-2573922549 (thanks @dimbleby 🙏), Poetry 1.8 can solve it in 0.2 s, no matter if there is only one extra or 14 extras. Poetry 2 takes longer the more extras are defined:

```toml
[tool.poetry.dependencies]
python = ">=3.8,<3.11"
opencv-python = {version = "^4.6.0.66", optional = true}
ale-py = { version = "0.8.1", optional = true }

[tool.poetry.extras]
extra01 = ["ale-py"]
extra02 = ["ale-py"]
extra03 = ["ale-py"]
# ...
```

number of extras|poetry 2.0|PR
---|---|---
1|6.5 s|0.5 s
2|>100 s|1.7 s
3|>100 s|6.4 s
4|>100 s|19 s
5|>100 s|57 s

## Summary by Sourcery

Tests:
- Updated tests for the solver internals to reflect the performance improvements.